+++ RELEASENOTES.md
@@ -139,6 +139,7 @@
         ([#1055](https://github.com/androidx/media/pull/1055)).
     *   Maintain a consistent luminance range across different HDR content (uses
         the HLG range).
+    *   Add support for Ultra HDR (bitmap) overlays on HDR content.
 *   Muxers:
 *   IMA extension:
     *   Promote API that is required for apps to play